requirements.txt
streamlit
keras 
matplotlib
seaborn
Pillow
tensorflow
opencv-python-headless
protobuf==3.9.2